+++ custom_components/panda_esl/sensor.py
@@ -19,7 +19,7 @@ from homeassistant.helpers.entity import EntityCategory
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 from homeassistant.helpers.update_coordinator import CoordinatorEntity
 
-from .const import DOMAIN, MANUFACTURER, MODEL
+from .const import DOMAIN, MANUFACTURER
 from .models import PandaEslState
 from .runtime import PandaEslRuntimeData
 
@@ -84,7 +84,7 @@ class PandaEslWriteProgressSensor(CoordinatorEntity, SensorEntity):
             connections={(CONNECTION_BLUETOOTH, self._runtime.state.address)},
             identifiers={(DOMAIN, self._runtime.state.address)},
             manufacturer=MANUFACTURER,
-            model=MODEL,
+            model=self._runtime.profile.model,
             name=self._entry.data.get(CONF_NAME) or self._entry.title,
         )
 
@@ -130,7 +130,7 @@ class PandaEslBluetoothRssiSensor(CoordinatorEntity, SensorEntity):
             connections={(CONNECTION_BLUETOOTH, self._runtime.state.address)},
             identifiers={(DOMAIN, self._runtime.state.address)},
             manufacturer=MANUFACTURER,
-            model=MODEL,
+            model=self._runtime.profile.model,
             name=self._entry.data.get(CONF_NAME) or self._entry.title,
         )